Cleaned up playerController, deleted duplicate playerControllers
