--- Assets/Scripts/JacobPlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6bef06b99e0291f4f852365f8f6e789f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
