--- Assets/Scripts/JerryPlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e0f03638e6def6e4aa90ed2f2acde736
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
