+++ Assets/Scripts/PlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f514a698c7121c048976718a36f66d45
+guid: 6bef06b99e0291f4f852365f8f6e789f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
